% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{FastQC}
\alias{FastQC}
\title{Construct FastQC object}
\usage{
FastQC(fqc)
}
\arguments{
\item{fqc}{A \code{list} of test data such as that returned by \link{parse_fqc}}
}
\value{
An object of class FastQC
}
\description{
Construct FastQC object
}
