% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{multiFastQC}
\alias{multiFastQC}
\title{Construc multiFastQC object}
\usage{
multiFastQC(fqc)
}
\arguments{
\item{fqc}{A \code{list} of test data such as that returned by \link{parse_fqc}}
}
\value{
An object of class multiFastQC
}
\description{
Construc multiFastQC object
}
