% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_collection}
\alias{parse_collection}
\title{Parse fastQC collection of files}
\usage{
parse_collection(file_names, names = NULL)
}
\arguments{
\item{file_names}{A \code{character} vector of file names}

\item{names}{A \code{character} vector of length equals that of file_names}
}
\value{
A \code{list} os lists.
}
\description{
Parse fastQC collection of files
}
