% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_data}
\alias{parse_data}
\title{Parse test data.frames}
\usage{
parse_data(txt, start, end)
}
\arguments{
\item{txt}{A \code{character} string such as this returned by \code{\link[base]{readLines}}}

\item{start}{An \code{integer} to mark the beginning of the table for a test}

\item{end}{An \code{integer} to mark the end of the table for a test}
}
\value{
A \code{data.frame} of the test data.
}
\description{
Parse test data.frames
}
