% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_fqc}
\alias{parse_fqc}
\title{Parse fastQC file}
\usage{
parse_fqc(file_name)
}
\arguments{
\item{file_name}{A \code{character} refers to the text file from fastQC.}
}
\value{
A named \code{list} containing a \code{data.frame} for each test.
}
\description{
A wrapper function to fully parse the text file output of FastQC in tidy
data.frames. Breifly, the function reads the text lines, parse the test
names, parse the test blocks of text and returns a list.
}
\examples{
fl <- system.file("extdata", "sample_1.txt", package = "Rfastqc")
fqc <- parse_fqc(fl)

}
