% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_test}
\alias{parse_test}
\title{Parse test output}
\usage{
parse_test(txt)
}
\arguments{
\item{txt}{A \code{character} string such as this returned by \code{\link[base]{readLines}}}
}
\value{
A \code{data.frame}
}
\description{
Parse test output
}
