% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot}
\alias{plot}
\title{Plot method for FastQC objects}
\usage{
plot(fqc, test = NULL, ...)
}
\arguments{
\item{fqc}{An object of class FastQC such as that returned by \link{FastQC}}

\item{test}{A \code{character} string of the test name}

\item{...}{Other arguments to pass to \code{link[graphics]{plot}}}
}
\value{
A line graph.
}
\description{
Plot method for FastQC objects
}
