% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot_tst}
\alias{plot_tst}
\title{Plot test data}
\usage{
plot_tst(tst_dat, test, ...)
}
\arguments{
\item{tst_dat}{A \code{data.frame} containing data for a single test.}

\item{test}{A \code{character} string with the test name.}

\item{...}{Other arguments passed to the \code{\link[graphics]{plot}} function.}
}
\value{
A line graph
}
\description{
Plot test data
}
