% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary}
\alias{summary}
\title{Summary method for FastQC objects}
\usage{
summary(fqc)
}
\arguments{
\item{fqc}{An object of class FastQC such as that returned by \link{FastQC}}
}
\value{
Prints summary of test data
}
\description{
Summary method for FastQC objects
}
