% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{test_plot}
\alias{test_plot}
\title{Plot test data}
\usage{
test_plot(tst_dat, test, color = "black", ...)
}
\arguments{
\item{tst_dat}{A \code{data.frame} containing data for a single test.}

\item{test}{A \code{character} string of the test name.}

\item{color}{A \code{character} vector of length equals to the number of
lines to be plotted.}

\item{...}{Other arguments passed to the \code{\link[graphics]{plot}}
function.}
}
\value{
A line graph
}
\description{
The function generates line graphs of the different test data. Mainly, its
a wrapper around the base R \code{\link[graphics]{plot}} to handle the
different types of tests.
}
