% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{multiFastQC}
\alias{multiFastQC}
\title{Construct multiFastQC object}
\usage{
multiFastQC(fqc)
}
\arguments{
\item{fqc}{A \code{list} of test data such as that returned by \link{parse_fqc}}
}
\value{
An object of class \code{multiFastQC}
}
\description{
Construct multiFastQC object
}
\examples{
path <- system.file('extdata', package = 'Rfastqc')
fls <- list.files(path, full.names = TRUE)
fqc <- parse_collection(fls)
fqc_ob <- multiFastQC(fqc)

}
