% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_data}
\alias{parse_data}
\title{Parse test data.frames}
\usage{
parse_data(txt, start, end)
}
\arguments{
\item{txt}{A \code{character} string such as this returned by
\code{\link[base]{readLines}}}

\item{start}{An \code{integer} to mark the beginning of the table for a test}

\item{end}{An \code{integer} to mark the end of the table for a test}
}
\value{
A \code{data.frame} of the test data.
}
\description{
The function parses the block of text of a particular test and returns a
tidy data.frame. The block of relevant text is first identified by a start
and and end line numbers and parsed as strings which is split later and
made a data.frame of.
}
\examples{
fl <- system.file("extdata", "sample_1.txt", package = "Rfastqc")
txt <- readLines(fl)
tst <- parse_test(txt)
ind <- parse_index(txt, tst)
dat <- parse_data(txt, start = ind$start[1], ind$end[1])

}
