% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_fqc}
\alias{parse_fqc}
\title{Parse fastQC file}
\usage{
parse_fqc(file_name)
}
\arguments{
\item{file_name}{A \code{character} refers to the text file from fastQC.}
}
\value{
A \code{list} containing a \code{data.frame} for each test
}
\description{
Parse fastQC file
}
