% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_index}
\alias{parse_index}
\title{Parse indecies for test data.frames}
\usage{
parse_index(txt, tst)
}
\arguments{
\item{txt}{A \code{character} string such as this returned by \code{\link[base]{readLines}}}

\item{tst}{A \code{data.frame} with test output summary such as that returned by \link{parse_test}}
}
\value{
A \code{data.frame} with the indices of the tests
}
\description{
Parse indecies for test data.frames
}
