% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_test}
\alias{parse_test}
\title{Parse test output}
\usage{
parse_test(txt)
}
\arguments{
\item{txt}{A \code{character} string such as this returned by
\code{\link[base]{readLines}}}
}
\value{
A \code{data.frame} of two columns \code{name} and \code{output}
fot the test name and output in terms of pass or fail.
}
\description{
The function parses the different test names and output in terms of pass or
fail from the text file output of the FastQC utility. Baically, the
function identies the test names lines' preceeded by '>>', parses the text
name and the output of that test in a \code{data.frame}.
}
\examples{
fl <- system.file("extdata", "sample_1.txt", package = "Rfastqc")
txt <- readLines(fl)
tst <- parse_test(txt)

}
