% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot}
\alias{plot}
\title{Plot method for FastQC/multiFastQC objects}
\usage{
plot(fqc, test = NULL, ...)
}
\arguments{
\item{fqc}{An object of class FastQC such as that returned by \link{FastQC}
or \link{multiFastQC}}

\item{test}{A \code{character} string of the test name.}

\item{...}{Other arguments to pass to \code{\link[graphics]{plot}}}
}
\value{
A line graph.
}
\description{
Plot method for FastQC/multiFastQC objects
}
\examples{
fl <- system.file("extdata", "sample_1.txt", package = "Rfastqc")
fqc <- parse_fqc(fl)
fqc <- FastQC(fqc)
plot(fqc, test = 'Per base sequence quality')

}
