% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{test_get}
\alias{test_get}
\title{Extract data for a particular test}
\usage{
test_get(fqc, test = NULL)
}
\arguments{
\item{fqc}{An object of class FastQC such as that returned by \link{FastQC}
or \link{multiFastQC}}

\item{test}{A \code{character} string of the test name.}
}
\value{
A \code{data.frame} of the test data.
}
\description{
Extract the data of a particular test from a \code{FastQC} or
\code{multiFasQC} object.
}
\examples{
fl <- system.file("extdata", "sample_1.txt", package = "Rfastqc")
fqc <- parse_fqc(fl)
fqc <- FastQC(fqc)
tst_dat <- test_get(fqc, test = 'Per base sequence quality')

}
